+++ src/Tests/Utilities/FakeRepository_test.hpp
@@ -53,7 +53,7 @@ namespace Allocation::Tests
         /// @brief Обновляет продукт.
         /// @param product Продукт для добавления.
         /// @note Реализовано для работы с TrackingRepository.
-        virtual void Update(Domain::ProductPtr product, int) override
+        virtual void Update(Domain::ProductPtr product, size_t) override
         {
             _skuByProduct.insert_or_assign(product->GetSKU(), product);
         }
